Added gitignore for local.properties in android root. Also beginning to move components around for better structure
